Split up scheduler.c to increase understandibility Everything that has to do with the ring_buffer is now in scheduler.c, everything else (dealing with tasks) is now in tasks.c
