+++ Makefile.am
@@ -30,6 +30,7 @@ noinst_PROGRAMS = ninjastorms
 ninjastorms_SOURCES = \
     kernel/boot/start.S \
     kernel/main.c kernel/main.h \
+    kernel/tasks.c kernel/tasks.h \
     kernel/scheduler.c kernel/scheduler.h \
     kernel/interrupt.c kernel/interrupt.h \
     kernel/syscall_handler.c kernel/syscall_handler.h \
